Update auto splitter for newest version
The newest version of My Friend Pedro changed the in-game time array to only update once a level is beaten. This is an issue because the auto splitter cannot sync to the in-game time with the old method any longer.
The best way to counter this is to simply make the same calculations that the `UITimerScript` class does (`Time.unscaledTime - root.timePaused - root.startTime`), but finding `TimeManager` is a time-consuming and non-trivial process.
So for now, we simply parse the display time from its string (`0:00.<size=9>00</size>`).